move metrics and avoid merge conflicts with epri metrics work
